% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_component.R
\name{ash_component}
\alias{ash_component}
\title{Ash component distribution}
\usage{
ash_component(ash_args = list())
}
\arguments{
\item{ash_args}{a list of params that get passed to \code{ashr::ash}}
}
\description{
Initialize an ash component distribution
when the distribution is updated, it initialized the new ash fit with the estimated \code{g} from the last iterations
this has the effect of only updating the mixture weights, but no the mixture distribution.
}
\details{
By default we use a normal mixture distribution for g (note: no point mass at 0)
}
