% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{autoselect_scales}
\alias{autoselect_scales}
\title{adapted from autoselect.mxisqp
try to select a default range for the sigmaa values
that should be used, based on the values of betahat and sebetahat
mode is the location about which inference is going to be centered
gridmult is the multiplier by which the sds differ across the grid}
\usage{
autoselect_scales(betahat, sebetahat, max_class, mult = 2)
}
\description{
adapted from autoselect.mxisqp
try to select a default range for the sigmaa values
that should be used, based on the values of betahat and sebetahat
mode is the location about which inference is going to be centered
gridmult is the multiplier by which the sds differ across the grid
}
