% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_moococomo_obj.R
\name{cal_ind_moment12}
\alias{cal_ind_moment12}
\title{Compute individual posterior first and second moment}
\usage{
cal_ind_moment12(fit, data, t_post_var, i)
}
\arguments{
\item{fit}{a mococomo object}

\item{i}{individual of interest}

\item{t_ind_var}{output of t_ind_var (using same mocomo object!)}
}
\description{
Compute individual posterior first and second moment
}
\examples{
t_post_var <-   do.call(rbind,
                       lapply( 1:length(data$betahat),
                               function(i)t_ind_var.mococomo(fit, i)
                       )
)


post_beta <-     do.call(c, lapply( 1: length(data$betahat), function(i)cal_ind_postmean(fit, t_post_var,i,) ))
}
