% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_computation.R
\name{cal_ind_moment12_mix_norm}
\alias{cal_ind_moment12_mix_norm}
\title{Compute individual posterior first and second moment}
\usage{
cal_ind_moment12_mix_norm(fit, data, t_post_var, i)
}
\arguments{
\item{fit}{a como object}

\item{i}{individual of interest}

\item{t_ind_var}{output of t_ind_var (using same mocomo object!)}
}
\description{
Compute individual posterior first and second moment

# TODO: currently only for center prior
}
