% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cEBMF.R
\name{cEBMF}
\alias{cEBMF}
\title{covariate moderated Empirical Bayes Matrix Factorization}
\usage{
cEBMF(
  Y,
  X_l,
  X_f,
  mnreg_type.x = "nnet",
  mnreg_type.y = "nnet",
  K = 1,
  type_noise = "constant",
  init_type = "udv_si",
  maxit = 100,
  tol = 0.001,
  param_como.x = list(max_class = 10, mnreg_type = "mult_reg"),
  param_como.y = list(max_class = 10, mnreg_type = "mult_reg"),
  param_nnet.x = list(size = 1, decay = 1),
  param_nnet.y = list(size = 1, decay = 1),
  param_como2 = list(),
  param_susie = list(L = 5),
  maxit_como = 10,
  dynamic_cov = NULL,
  check_l_prior = NULL,
  intercept = FALSE
)
}
\arguments{
\item{Y}{numerical matrix size NxP}

\item{X_l}{matrix of size NxJ containning covariates affecting the factors}

\item{X_f}{matrix of size PxT containning covariates affecting the factors}

\item{K}{numeric number of factors}

\item{type_noise}{specify which kind of noise structure is expected, currently three choices. Whether noise constant accross column ('column_wise'), constant 'constant' or constant across rown 'row_wise'}

\item{init_type}{specify the initialization method, the following methods are available 'udv_si', 'random', 'flashier', 'flashier_NMF', 'flashier_SNM', 'udv_si_svd',
'udv','rand_udv'. Default is udv_si.}

\item{maxit}{maximum nuber of iterations}

\item{tol}{paramter for assessing convergence}

\item{param_como.x}{list of parameters for como object for X_l}

\item{param_como.y}{list of parameters for como object for Y}

\item{param_nnet.x}{list of parameters for nnet object for X_f}

\item{param_nnet.y}{list of parameters for nnet object for Y}

\item{param_como2}{list of parameters for como object for Y}

\item{param_susie}{list of parameters for susie object for Y}

\item{maxit_como}{maximum number of iterations for como object}

\item{mnreg_type}{specify underlying learner for within topic heterogeneity , two methods available 'nnet' or 'keras'. Default is nnet.
you can pass the nnet specification through the param_nnet argument and through the param_susie for logistic susie.}
}
\value{
a cEBMF object
}
\description{
covariate moderated Empirical Bayes Matrix Factorization
}
