% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{cFDR}
\alias{cFDR}
\title{Main FDR function}
\usage{
cFDR(
  betahat,
  se,
  pvalue,
  zscore,
  X,
  coverage = 0.95,
  maxiter = 100,
  tol = 0.001,
  max_class = 10,
  mult = 2,
  upper = FALSE,
  outputlevel = 1,
  n_sim = 1000,
  alpha,
  nullweight = 4,
  verbose = TRUE
)
}
\arguments{
\item{outputlevel}{Determines amount of output. outputlevel=1 provide simplest output (which should be enough for most applications)
outputlevel= also output the entire como fitted object}
}
\description{
Main function gathering all the routines for different FDR estimation procedures
}
\details{
if the user specifies (betahat, se) then cFDR will fit an covariate
moderated ash model. If the user specifies p-value, then cFDR fits a mixture of betas
which only contains distribution with decreasing density (mostly focusing on fitting the leftmost tails).
If the user specifies upper=TRUE when using the pvalue argument then cFDR fits a mixture of betas
which  contains distribution with decreasing density or distribution with increasing density (thus fitting both tail of the distribution).
If the user specifies the model essentially fit a extension of the ZAP procedure (projected zscore of 0-1) which
corresponds to fitting the same model as for pvalue argument with upper=TRUE;  (mixture of betas
which  contains distribution with decreasing density or distribution with increasing density)
}
