% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticsusie-package.R
\docType{package}
\name{comoR-package}
\alias{comoR}
\alias{comoR-package}
\title{comoR: covariate moderated FDR methods based SuSiE}
\description{
Implements logistic, multinomial and categorical SuSiE as well as some additional FDR and shrinkage methods
}
\author{
\strong{Maintainer}: Karl Tayeb \email{ktayeb@uchicago.edu}

Authors:
\itemize{
  \item William  Denault \email{wdenault@uchicago.edu}
}

}
\keyword{internal}
