% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_regression.R
\name{compute_log_prior_assignment}
\alias{compute_log_prior_assignment}
\alias{compute_log_prior_assignment.mult_reg}
\alias{compute_log_prior_assignment.keras_obj}
\title{Compute the prior probability of assignment to each mixture component}
\usage{
compute_log_prior_assignment(x, ...)

\method{compute_log_prior_assignment}{mult_reg}(mnreg, data)

\method{compute_log_prior_assignment}{keras_obj}(mnreg, data)
}
\description{
Compute the prior probability of assignment to each mixture component
}
\keyword{internal}
