% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{compute_prior_log_odds}
\alias{compute_prior_log_odds}
\title{Compute the prior probability of assignment to each mixture component}
\usage{
compute_prior_log_odds(x, ...)
}
\description{
Compute the prior probability of assignment to each mixture component
}
