% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{convolved_logpdf.exp}
\alias{convolved_logpdf.exp}
\title{Convolution between  an exponential and Normal distribution with variance given by sum of component dist. and noise}
\usage{
\method{convolved_logpdf}{exp}(dist, betahat, se)
}
\description{
Convolution between  an exponential and Normal distribution with variance given by sum of component dist. and noise
}
