% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{convolved_logpdf.normal}
\alias{convolved_logpdf.normal}
\title{Normal distribution with variance given by sum of component dist. and noise}
\usage{
\method{convolved_logpdf}{normal}(dist, betahat, se)
}
\description{
Normal distribution with variance given by sum of component dist. and noise
}
