% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{convolved_logpdf.point}
\alias{convolved_logpdf.point}
\title{Just a normal distribution with mean centered at the point mass}
\usage{
\method{convolved_logpdf}{point}(dist, betahat, se)
}
\description{
Just a normal distribution with mean centered at the point mass
}
