% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/como.R
\name{data_initialize_como}
\alias{data_initialize_como}
\title{data_initialize_como
@param data description
@param scales description
@param mu0 description
@param var0 description
@param nullweight description

@param mnreg_type description
@param param_nnet description}
\usage{
data_initialize_como(
  data,
  prior = c("mix_norm", "mix_exp"),
  max_class = 10,
  scales = NULL,
  mu0 = 0,
  var0 = 1,
  nullweight = 0,
  mnreg_type = "mnreg_constant",
  param_nnet = list(size = 1, decay = 1),
  epoch = 10,
  verbose_keras = 0,
  batch_size = 100,
  weights = NULL
)
}
\description{
Use data to autoselect scales
}
