% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mococomo.R
\name{fit.mococomo}
\alias{fit.mococomo}
\title{Function implementation the mococomo mode}
\usage{
fit.mococomo(
  data,
  model = "normal",
  maxiter = 100,
  tol = 0.001,
  max_class = 10,
  mult = 2,
  upper = FALSE,
  nullweight
)
}
\arguments{
\item{data}{an object of class data_mococomo  see \link{\code{set_data_mococomo}}}

\item{maxiter}{numeric, maximum numerous of iteration set to 100 by defaults}

\item{tol}{tolerance in term of change in ELBO value for stopping criterion}

\item{upper, }{logical, set to FALSE by default. Specific to beta distribution.
If true use a to set of mixture for fitting both end of the of the distribution as in the ZAP paper by Leung and Sunn}

\item{modeltype}{of model currently supported (normal and beta )}
}
\description{
Function implementation the mococomo mode
}
\details{
Function implementation the mococomo mode
}
