% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_computation.R
\name{get_fdr.como}
\alias{get_fdr.como}
\title{Compute individual fdr value  como model with centered normal mixture}
\usage{
\method{get_fdr}{como}(fit, data)
}
\arguments{
\item{fit}{a como object}
}
\description{
Compute individual fdr value  como model with centered normal mixture
}
