% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_fdr}
\alias{get_fdr}
\title{Compute individual fdr value como/como2 model with centered normal mixture}
\usage{
get_fdr(fit, data, i, ...)
}
\description{
Compute individual fdr value como/como2 model with centered normal mixture
}
