% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_como2.R
\name{get_lfdr.como2}
\alias{get_lfdr.como2}
\title{Compute individual fdr value   como2 model with centered normal mixture}
\usage{
\method{get_lfdr}{como2}(fit, data)
}
\arguments{
\item{fit}{a como object}
}
\description{
Compute individual fdr value   como2 model with centered normal mixture
}
