% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_lfdr}
\alias{get_lfdr}
\title{Compute individual lfdr value como/como2 model with centered normal mixture}
\usage{
get_lfdr(fit, data, i, ...)
}
\description{
Compute individual lfdr value como/como2 model with centered normal mixture
}
