% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{initialize_linear_susie}
\alias{initialize_linear_susie}
\title{Initialize linear SuSiE logistic regression object}
\usage{
initialize_linear_susie(n, ...)
}
\arguments{
\item{n}{number of observations}
}
\description{
dots are captured in a list and passed as arguments to susieR::susie
}
