% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{initialize_logistic_ibss}
\alias{initialize_logistic_ibss}
\title{Initialize linear SuSiE logistic regression object}
\usage{
initialize_logistic_ibss(n, ...)
}
\arguments{
\item{n}{number of observations}
}
\description{
dots are captured in a list and passed as arguments to susieR::susie
}
