% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_como2.R
\name{post_mean_sd.como2}
\alias{post_mean_sd.como2}
\title{Compute individual posterior mean and sd under como model}
\usage{
\method{post_mean_sd}{como2}(fit, data)
}
\arguments{
\item{fit}{a como object}
}
\description{
Compute individual posterior mean and sd under como model

# TODO: allow using new observation from another data set (e.g. testing)
}
