% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{post_mean_sd}
\alias{post_mean_sd}
\title{Compute post first and second moment from como and como2 object}
\usage{
post_mean_sd(fit, data, ...)
}
\description{
Compute post first and second moment from como and como2 object
}
