% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit_model}
\alias{update_model}
\alias{compute_elbo}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{logisticsusie}{\code{\link[logisticsusie]{compute_elbo}}, \code{\link[logisticsusie]{fit_model}}, \code{\link[logisticsusie]{update_model}}}
}}

