% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_computation.R
\name{t_ind_var.como}
\alias{t_ind_var.como}
\title{Compute individual posterior variance from marginal normal mean model}
\usage{
\method{t_ind_var}{como}(fit, data, i)
}
\description{
internal function to compute posterior mean and sds
}
