% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_como2.R
\name{t_ind_var.como2}
\alias{t_ind_var.como2}
\title{Compute individual posterior variance from marginal normal mean model}
\usage{
\method{t_ind_var}{como2}(fit, data, i)
}
\description{
internal function to compute posterior mean and sds
}
