% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{t_ind_var}
\alias{t_ind_var}
\title{Compute individual posterior variance from marginal normal mean model}
\usage{
t_ind_var(fit, data, i, ...)
}
\description{
internal function to compute posterior mean and sds
}
