% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{update_params.normal}
\alias{update_params.normal}
\title{Note: only updates the variance parameter, these components are assumed
to be mean 0}
\usage{
\method{update_params}{normal}(dist, betahat, se, weights)
}
\description{
Note: only updates the variance parameter, these components are assumed
to be mean 0
}
