% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cEBMF.R
\name{cEBMF}
\alias{cEBMF}
\title{covariate moderated Empirical Bayes Matrix Factorization}
\usage{
cEBMF(
  Y,
  X_l,
  X_f,
  mnreg_type = "nnet",
  K = 1,
  type_noise = "constant",
  init_type = "udv_si",
  maxit = 100,
  tol = 0.001,
  param_como.x = list(max_class = 10, mnreg_type = "mult_reg"),
  param_como.y = list(max_class = 10, mnreg_type = "mult_reg"),
  param_nnet.x = list(size = 1, decay = 1),
  param_nnet.y = list(size = 1, decay = 1),
  param_como2 = list(),
  param_susie = list(L = 5),
  maxit_como = 10
)
}
\arguments{
\item{Y}{numerical matrix size NxP}

\item{X_l}{matrix of size NxJ containning covariates affecting the factors}

\item{X_f}{matrix of size PxT containning covariates affecting the factors}

\item{mnreg_type}{specify underlying learner for within topic heterogeneity , two methods available 'nnet' or 'logistic_susie'. Default is nnet.
you can pass the nnet specification through the param_nnet argument and through the param_susie for logistic susie.}

\item{K}{numeric number of factors}

\item{type_noise}{specify which kind of noise structure is expected, currently three choices. Whether noise constant accross column ('column_wise'), constant 'constant' or constant across rown 'row_wise'}

\item{maxit}{maximum nuber of iterations}

\item{tol}{paramter for assessing convergence}
}
\value{
a cEBMF object
}
\description{
covariate moderated Empirical Bayes Matrix Factorization
}
