% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{como_prep_data}
\alias{como_prep_data}
\title{Prepare data for multinomial regression}
\usage{
como_prep_data(betahat, se, X, Z = NULL, center = TRUE, scale = FALSE)
}
\arguments{
\item{betahat}{vector of effect estimates}

\item{se}{vector of standard errors for effect estimates}

\item{X}{matrix of features, each row an observation}

\item{Z}{matrix of fixed-effect covariates (including intercept), each row an observation}

\item{center}{boolean to center \code{X}}

\item{scale}{boolean to scale \code{X} to unit variance}
}
\description{
Prepare data for multinomial regression
}
