% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_regression.R
\name{compute_log_prior_assignment}
\alias{compute_log_prior_assignment}
\title{Compute the prior probability of assignment to each mixture component}
\usage{
compute_log_prior_assignment(x, ...)
}
\description{
Compute the prior probability of assignment to each mixture component
}
