% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/como.R
\name{data_initialize_como}
\alias{data_initialize_como}
\title{Use data to autoselect scales}
\usage{
data_initialize_como(
  data,
  max_class,
  scales = NULL,
  mu0 = 0,
  var0 = 1,
  nullweight = 0,
  mnreg = "constant",
  param_nnet = list(size = 1, decay = 1)
)
}
\description{
Use data to autoselect scales
}
