% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_como_obj.R
\name{get_fdr}
\alias{get_fdr}
\title{Compute individual fdr value  como model with centered normal mixture}
\usage{
get_fdr(fit)
}
\arguments{
\item{fit}{a como object}
}
\description{
Compute individual fdr value  como model with centered normal mixture
}
